+++ services/src/main/java/ru/dpopkov/knownics/dto/KeyTermDto.java
@@ -1,0 +1,14 @@
+package ru.dpopkov.knownics.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.validation.constraints.NotEmpty;
+
+@Getter
+@Setter
+public class KeyTermDto extends ModifiableDto {
+    @NotEmpty
+    private String name;
+    private String description;
+}