+++ services/src/main/java/ru/dpopkov/knownics/dto/converters/DtoToQuestionText.java
@@ -1,0 +1,15 @@
+package ru.dpopkov.knownics.dto.converters;
+
+import org.springframework.core.convert.converter.Converter;
+import org.springframework.stereotype.Component;
+import ru.dpopkov.knownics.domain.question.QuestionText;
+import ru.dpopkov.knownics.dto.TranslationDto;
+
+@Component
+public class DtoToQuestionText extends DtoToTranslation<QuestionText>
+                               implements Converter<TranslationDto, QuestionText> {
+
+    public DtoToQuestionText() {
+        super(QuestionText::new);
+    }
+}